buildsteps_feed.py: add "alwaysUseLatest" to git-step
The option alwaysUseLatest for gitsteps ignores the commit hash that was
delivered with the webhook-shot. Instead, it always uses the latest commit
of the defined branch.
We need to ignore the commithashes, as they match commits from the packages
repository, but we want to use the builter-repo to build them.

Signed-off-by: Martin Hübner <martin.hubner@web.de>